requirements.txt
asgiref==3.8.1
Django==4.2.11
docx2txt==0.8
et-xmlfile==1.1.0
lxml==5.2.1
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
pypdf2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
typing-extensions==4.11.0
tzdata==2024.1
